Operators can now run a single command to implants and receive output from them
